requirements.txt
numpy
scipy>=1.3
sympy>=1.5
tqdm
ase
numba
python==3.8
jupyterlab
pytorch==1.9.0
torchvision 
torchaudio 
cudatoolkit=10.2
pytorch-geometric
easydict
